vision/requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
msrest>=0.7.1
Pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.21.0 